# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mission Control API",
    "version": "1.2.0",
    "description": "AI Agent Orchestration Platform API"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    },
    {
      "apiKey": []
    }
  ],
...
  "paths": {
...
    "/api/agents/{id}/heartbeat": {
      "get": {
        "tags": [
          "Agents"
        ],
        "summary": "Check agent work items",
        "operationId": "getAgentHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Heartbeat data with pending work items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "HEARTBEAT_OK",
                        "WORK_ITEMS_FOUND"
                      ]
                    },
                    "agent": {
                      "type": "string"
                    },
                    "checked_at": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "assigned_tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TaskSummary"
                      }
                    },
                    "in_progress_tasks": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/TaskSummary"
                      }
                    },
                    "has_actionable_work": {
                      "type": "boolean"
                    },
                    "work_items": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "total_items": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "status",
                    "agent",
                    "checked_at",
                    "assigned_tasks",
                    "in_progress_tasks",
                    "has_actionable_work"
                  ]
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "post": {
        "tags": [
          "Agents"
        ],
        "summary": "Trigger manual heartbeat",
        "operationId": "triggerAgentHeartbeat",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Heartbeat triggered",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mc-session"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TaskSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "priority": {
            "type": "string"
          },
          "due_date": {
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "priority"
        ]
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    }
  }
}